Filter: Improve handling of sets in BGP path masks
This is a follow-up for commit ef113c6f7253 ("Filter: Allow to use sets
in path masks", 2019-08-06).

Compare the content of PM_ASN_SET in path masks. A reconfiguration
was not properly triggering a reload of affected protocols when the
members of a set in a path mask change.

Also, update the printing code to so that it can display sets in a path
mask.