Netlink: Allocate static sized buffer for netlink message
The current logic to calculate the required buffer size to construct a
netlink message is erroneous in many ways.

Use a static sized buffer to simplify things.

FIXME: this change is not complete and has a serious bug - if the
constructed message happens to be larger than 8192 bytes, BIRD will
hit bug() and crash.